chore: move files to auth feature dir
